Object orientation detection now working
Implimented a method to detect the orientation of object relative to the target transfor marker. Needs additional testing and refinement with larger board, however, basic testing now shows as working. Also implimented an array to record when objects are detected as in place. Next development step is to ray test the objects and nodes to ensure that their placement is correct. A good and producte day!